Enforced api services with checks of authorizations

+++ src/EthernaIndex/Areas/Api/Services/VideosControllerService.cs
@@ -3,10 +3,12 @@ using Etherna.EthernaIndex.Areas.Api.InputModels;
 using Etherna.EthernaIndex.Domain;
 using Etherna.EthernaIndex.Domain.Models;
 using Etherna.EthernaIndex.Domain.Models.Swarm;
+using Etherna.EthernaIndex.Extensions;
 using Etherna.MongODM.Extensions;
-using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Http;
 using MongoDB.Driver;
 using MongoDB.Driver.Linq;
+using Nethereum.Util;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -17,18 +19,23 @@ namespace Etherna.EthernaIndex.Areas.Api.Services
     internal class VideosControllerService : IVideosControllerService
     {
         // Fields.
+        private readonly IHttpContextAccessor httpContextAccessor;
         private readonly IIndexContext indexContext;
 
         // Constructors.
-        public VideosControllerService(IIndexContext indexContext)
+        public VideosControllerService(
+            IHttpContextAccessor httpContextAccessor,
+            IIndexContext indexContext)
         {
+            this.httpContextAccessor = httpContextAccessor;
             this.indexContext = indexContext;
         }
 
         // Methods.
-        public async Task<VideoDto> CreateAsync(string channelAddress, VideoCreateInput videoInput)
+        public async Task<VideoDto> CreateAsync(VideoCreateInput videoInput)
         {
-            var channel = await indexContext.Channels.FindOneAsync(c => c.Address == channelAddress);
+            var address = httpContextAccessor.HttpContext.User.GetEtherAddress();
+            var channel = await indexContext.Channels.FindOneAsync(c => c.Address == address);
             var manifestHash = new SwarmContentHash(videoInput.ManifestHash);
 
             var video = new Video(
@@ -42,14 +49,19 @@ namespace Etherna.EthernaIndex.Areas.Api.Services
             return new VideoDto(video);
         }
 
-        public async Task<ActionResult> DeleteAsync(string hash)
+        public async Task DeleteAsync(string hash)
         {
+            // Get data.
+            var address = httpContextAccessor.HttpContext.User.GetEtherAddress();
             var video = await indexContext.Videos.QueryElementsAsync(elements =>
                 elements.FirstAsync(v => v.ManifestHash.Hash == hash));
 
-            await indexContext.Videos.DeleteAsync(video);
+            // Verify authz.
+            if (!video.OwnerChannel.Address.IsTheSameAddress(address))
+                throw new UnauthorizedAccessException("User is not owner of the video");
 
-            return new OkResult();
+            // Action.
+            await indexContext.Videos.DeleteAsync(video);
         }
 
         public async Task<VideoDto> FindByHashAsync(string hash) =>
@@ -63,10 +75,16 @@ namespace Etherna.EthernaIndex.Areas.Api.Services
 
         public async Task<VideoDto> UpdateAsync(string oldHash, string newHash)
         {
+            // Get data.
+            var address = httpContextAccessor.HttpContext.User.GetEtherAddress();
             var video = await indexContext.Videos.FindOneAsync(v => v.ManifestHash.Hash == oldHash);
 
-            video.SetManifestHash(new SwarmContentHash(newHash));
+            // Verify authz.
+            if (!video.OwnerChannel.Address.IsTheSameAddress(address))
+                throw new UnauthorizedAccessException("User is not owner of the video");
 
+            // Action.
+            video.SetManifestHash(new SwarmContentHash(newHash));
             await indexContext.SaveChangesAsync();
 
             return new VideoDto(video);
